+++ src/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/julesklord/bananascaler/cmd"
+
+func main() {
+	cmd.Execute()
+}